modifying init.sh

+++ init.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
-git clone git@github.com:hubdrop/cookbooks.git
-git clone git@github.com:hubdrop/app.git
+git submodule init
+git submodule update
 
 # @TODO: Let's make init.sh a wizard, and ask these things
-cp attributes.json.example attributes.json
+if [ ! -f /tmp/foo.txt ]; then
+  cp attributes.json.example attributes.json
+fi
+
+# @TODO: Only add to hosts if it isn't there already.
 echo '10.10.10.10  hubdrop.local' | sudo tee -a /etc/hosts
